backend/requirements.txt
async-timeout==4.0.2
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Reuploaded==1.3.0
Flask-SQLAlchemy==3.0.2
Flask-Uploads==0.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.6.0
python-dotenv==1.0.0
redis==4.6.0
schedule==1.2.0
SQLAlchemy==2.0.17
typing_extensions==4.7.0
Werkzeug==2.3.6
